base: Actually fix the padding on StatusBar
Android P gets a new QS layout that looks good but it has some issues on smaller screens devices, like the rounded buttons clipping with the borders and the date not looking really good.
LineageOS suggested a fix (https://review.lineageos.org/224021) but it's circumventing it because it adds padding and some people doesn't like the padding, so, instead let's apply the 8dp spacing over the affected headers and the small qs tile.

Android 10 edit: Let's add it back, but let's set the value to 0, because there's more notched devices that un-notched on our roster

Signed-off-by: spezi77 <spezi7713@gmx.net>