Crash app on foreground service notification error.
Resolved issue 118612296.

On any notification error, the NMS silently cancels the notification,
including foreground service notifications. Thus, an app could pass in a
garbage notification deliberately and start a foreground service silently.

This patch resolved this issue by judging the notification's flag, and if
it is a foreground notification, still crash the app as previous platforms,
and if it is a normal notification, don't crash the app.

Background: In 3ad4cdd1, which was merged into Android 9 release, the crash
behaviour is removed. But it is an important rule that foreground services
guaranteed to show an ongoing notification.

Test: Run the sample apk provided in the issue, it's main thread received
 a RemoteServiceException: Bad notification posted from package...
as intended behaviour.

Bug: 118612296
Merged-In: I94d9de50bb03c33666471e3dbd9c721e9278f7cb
Change-Id: Ice6faab055e2dacd3d2d12803fcf51e13d3b0f21
Signed-off-by: spezi77 <spezi7713@gmx.net>