SB clock font: Make it play nice with DU-Themes font overlay [1/2].
Let me explain why I did this.

This allows the user to use a specific font style for the status bar clock
in addition to a system-wide font style. As soon as the user changes back
to default font type in the Statusbar Settings, the system-wide font style
that was loaded via overlay may be used again.

Signed-off-by: spezi77 <spezi7713@gmx.net>